refactor: extract pagination into ApplicationListViewModel+Pagination to satisfy file_length (tc-ul1i)
