+++ mobile/ios/packages/town-crier-presentation/Sources/Features/ApplicationList/ApplicationListViewModel.swift
@@ -14,7 +14,10 @@ import TownCrierDomain
 /// single-select status-chip group.
 @MainActor
 public final class ApplicationListViewModel: ObservableObject, ErrorHandlingViewModel {
-  @Published private(set) var applications: [PlanningApplication] = []
+  // `internal(set)` (the default for this non-public property) rather than
+  // `private(set)`: the paged load/append lives in `+Pagination` extension file
+  // and mutates this. Still read-only outside the presentation module.
+  @Published var applications: [PlanningApplication] = []
   @Published var selectedStatusFilter: ApplicationStatus? {
     didSet {
       // Status and Unread chips share a single-select group (spec decision #7)
@@ -51,11 +54,13 @@ public final class ApplicationListViewModel: ObservableObject, ErrorHandlingView
     }
   }
 
-  private let repository: PlanningApplicationRepository?
-  private let offlineRepository: OfflineAwareRepository?
+  // `repository`/`offlineRepository`/`zone` are internal (not `private`) so the
+  // `+Pagination` extension file can drive the paged fetch.
+  let repository: PlanningApplicationRepository?
+  let offlineRepository: OfflineAwareRepository?
   private let watchZoneRepository: WatchZoneRepository?
   private let notificationStateRepository: NotificationStateRepository?
-  private var zone: WatchZone?
+  var zone: WatchZone?
   /// Re-entrancy guard for ``loadApplications()``. A single user action can
   /// trigger the load repeatedly — `.task` firing alongside `.refreshable`, a
   /// scenePhase change, or the view re-appearing — and each unguarded call
@@ -67,23 +72,23 @@ public final class ApplicationListViewModel: ObservableObject, ErrorHandlingView
   /// exist (the last page omits `X-Next-Cursor`) or the active sort is
   /// client-side. Reset to `nil` on every fresh first-page load and on any sort
   /// change so a stale cursor can never page a differently-ordered set
-  /// (GH#682 slice 1).
-  private var nextCursor: String?
+  /// (GH#682 slice 1). Internal so the `+Pagination` extension can drive it.
+  var nextCursor: String?
   /// The sort the currently-loaded `applications` were fetched under. Lets a
   /// sort change decide whether a refetch is needed (server sorts and any
   /// transition away from one reload; a client→client switch only re-sorts in
   /// memory, as before).
-  private var loadedSort: ApplicationsSort?
+  var loadedSort: ApplicationsSort?
   /// Re-entrancy guard for next-page fetches — independent of `isLoadInFlight`
   /// so an in-flight first-page load never blocks (or is blocked by) appends.
-  private var isPageLoadInFlight = false
+  var isPageLoadInFlight = false
   /// Server page size requested for the infinite-scroll path. Matches the API's
   /// sorted-path default (GH#682); the server clamps anything larger.
   static let pageSize = 150
   /// Trigger the next-page fetch once the appearing row is within this many rows
   /// of the end of the loaded set, so the next page lands before the user hits
   /// the bottom.
-  private static let prefetchThreshold = 10
+  static let prefetchThreshold = 10
   private let userDefaults: UserDefaults
   private let zoneSelectionKey: String
   private let sortKey: String
@@ -259,95 +264,6 @@ public final class ApplicationListViewModel: ObservableObject, ErrorHandlingView
     isLoading = false
   }
 
-  /// Loads the active zone's first page via the path appropriate to the current
-  /// sort. The three server-supported sorts (distance/newest/oldest) fetch a
-  /// server-ordered page and capture the next-page cursor; the client-side sorts
-  /// (recent-activity/status) keep the legacy param-less first-page fetch. Either
-  /// way the cursor resets, so pagination always restarts cleanly (GH#682).
-  private func loadActiveZone(_ zone: WatchZone) async throws {
-    if let order = sort.serverOrder {
-      let page = try await fetchPage(for: zone, sort: order, cursor: nil)
-      applications = page.applications
-      nextCursor = page.nextCursor
-    } else {
-      applications = try await fetchApplications(for: zone)
-      nextCursor = nil
-    }
-    loadedSort = sort
-  }
-
-  /// Fetches and appends the next server page when one exists. No-op unless the
-  /// active sort is server-driven, a cursor is held, and no append is already in
-  /// flight. Following `X-Next-Cursor` until it is absent walks the whole set;
-  /// the last page (no cursor) ends the loop. A fetch error surfaces to `error`.
-  public func loadNextPage() async {
-    guard let order = sort.serverOrder,
-      let cursor = nextCursor,
-      !isPageLoadInFlight,
-      let activeZone = selectedZone ?? zone
-    else { return }
-    isPageLoadInFlight = true
-    defer { isPageLoadInFlight = false }
-    do {
-      let page = try await fetchPage(for: activeZone, sort: order, cursor: cursor)
-      appendPage(page)
-    } catch {
-      handleError(error)
-    }
-  }
-
-  /// Called by the list as each row appears. Kicks off the next-page fetch when
-  /// a server sort is active, more pages remain, and the appearing row is within
-  /// ``prefetchThreshold`` of the end of the loaded set. A no-op for client-side
-  /// sorts, which hold the whole (first-page) set already.
-  public func onRowAppear(_ application: PlanningApplication) async {
-    guard sort.isServerSorted, nextCursor != nil, !isPageLoadInFlight else { return }
-    guard let index = applications.firstIndex(where: { $0.id == application.id }) else { return }
-    guard index >= applications.count - Self.prefetchThreshold else { return }
-    await loadNextPage()
-  }
-
-  /// Reacts to a sort change from the toolbar. A server sort — or any transition
-  /// away from one — reloads from page 1 with a fresh cursor so the list re-pages
-  /// in the new order. A switch between the two client-side sorts only changes
-  /// the in-memory ordering, so it skips the refetch (unchanged from the
-  /// pre-pagination behaviour). The cursor and `loadedSort` reset happen inside
-  /// the reload.
-  public func handleSortChanged() async {
-    guard sort != loadedSort else { return }
-    let newIsServer = sort.isServerSorted
-    let oldWasServer = loadedSort?.isServerSorted ?? false
-    if !newIsServer && !oldWasServer {
-      loadedSort = sort
-      return
-    }
-    await loadApplications()
-  }
-
-  /// Appends a server page, dropping any rows already loaded so a keyset overlap
-  /// at a page boundary never duplicates a row. Captures the new cursor.
-  private func appendPage(_ page: ApplicationPage) {
-    let existingIds = Set(applications.map(\.id))
-    applications.append(contentsOf: page.applications.filter { !existingIds.contains($0.id) })
-    nextCursor = page.nextCursor
-  }
-
-  private func fetchPage(
-    for zone: WatchZone,
-    sort order: ApplicationSortOrder,
-    cursor: String?
-  ) async throws -> ApplicationPage {
-    if let offlineRepository {
-      return try await offlineRepository.fetchApplicationsPage(
-        for: zone, sort: order, cursor: cursor, limit: Self.pageSize)
-    }
-    if let repository {
-      return try await repository.fetchApplicationsPage(
-        for: zone, sort: order, cursor: cursor, limit: Self.pageSize)
-    }
-    return ApplicationPage(applications: [], nextCursor: nil)
-  }
-
   public func selectApplication(_ id: PlanningApplicationId) {
     onApplicationSelected?(id)
   }
@@ -392,7 +308,9 @@ public final class ApplicationListViewModel: ObservableObject, ErrorHandlingView
     selectedZone = zones.first
   }
 
-  private func fetchApplications(for zone: WatchZone) async throws -> [PlanningApplication] {
+  /// Internal (not `private`): the `+Pagination` extension's `loadActiveZone`
+  /// uses this for the client-side sort path.
+  func fetchApplications(for zone: WatchZone) async throws -> [PlanningApplication] {
     if let offlineRepository {
       return try await offlineRepository.fetchApplications(for: zone).data
     } else if let repository {
